+++ src/c/transport/tcp.h
@@ -1,3 +1,4 @@
+#include "transport.h"
 #include <netinet/in.h>
 #include <stdint.h>
