+++ src/c/transport/transport.c
@@ -1,9 +1,9 @@
 #include "transport.h"
 
-void transport_init_base(struct rasta_handle *h, rasta_transport_channel* channel, unsigned id, const char *host, uint16_t port, const rasta_config_tls *tls_config) {
+void transport_init(struct rasta_handle *h, rasta_transport_channel *channel, unsigned id, const char *host, uint16_t port, const rasta_config_tls *tls_config) {
     channel->id = id;
     channel->remote_port = port;
-    strncpy(channel->remote_ip_address, host, INET_ADDRSTRLEN-1);
+    strncpy(channel->remote_ip_address, host, INET_ADDRSTRLEN - 1);
     channel->send_callback = send_callback;
     channel->tls_config = tls_config;
 
@@ -21,7 +21,7 @@ void transport_init_base(struct rasta_handle *h, rasta_transport_channel* channe
 
 // finds the transport channel corresponding to the sender (identified by IP address and port)
 // and returns the indexes of this transport channel and the redundancy channel it belongs to
-void find_channel_by_ip_address(struct rasta_handle *h, struct sockaddr_in sender, int *red_channel_idx, int *transport_channel_idx){
+void find_channel_by_ip_address(struct rasta_handle *h, struct sockaddr_in sender, int *red_channel_idx, int *transport_channel_idx) {
     *red_channel_idx = -1;
     *transport_channel_idx = -1;
 
@@ -31,8 +31,7 @@ void find_channel_by_ip_address(struct rasta_handle *h, struct sockaddr_in sende
     for (unsigned i = 0; i < h->mux.redundancy_channels_count; i++) {
         for (unsigned j = 0; j < h->mux.redundancy_channels[i].transport_channel_count; j++) {
             rasta_transport_channel *current_channel = &h->mux.redundancy_channels[i].transport_channels[j];
-            if (strncmp(current_channel->remote_ip_address, ip_addr, INET_ADDRSTRLEN) == 0
-                && current_channel->remote_port == ntohs(sender.sin_port)) {
+            if (strncmp(current_channel->remote_ip_address, ip_addr, INET_ADDRSTRLEN) == 0 && current_channel->remote_port == ntohs(sender.sin_port)) {
                 *red_channel_idx = i;
                 *transport_channel_idx = j;
                 break;
