+++ src/c/transport/transport.h
@@ -8,6 +8,8 @@
 
 #define MAX_PENDING_CONNECTIONS 5
 
+#define UNUSED(x) (void)(x)
+
 #ifdef ENABLE_TLS
 #include <wolfssl/options.h>
 #include <wolfssl/ssl.h>
@@ -39,7 +41,7 @@ typedef struct rasta_transport_channel {
 
     rasta_tls_mode tls_mode;
 
-    const rasta_config_tls * tls_config;
+    const rasta_config_tls *tls_config;
 
     /**
      * filedescriptor
@@ -91,7 +93,6 @@ typedef struct rasta_transport_socket {
 
 } rasta_transport_socket;
 
-
 void send_callback(redundancy_mux *mux, struct RastaByteArray data_to_send, rasta_transport_channel *channel, unsigned int channel_index);
 ssize_t receive_callback(struct receive_event_data *data, unsigned char *buffer, struct sockaddr_in *sender);
 
@@ -104,7 +105,5 @@ int transport_connect(rasta_transport_socket *socket, rasta_transport_channel *c
 int transport_redial(rasta_transport_channel *channel, rasta_transport_socket *socket);
 void transport_close(rasta_transport_channel *channel);
 
-
 // Protected methods
-void transport_init_base(struct rasta_handle *h, rasta_transport_channel *channel, unsigned id, const char *host, uint16_t port, const rasta_config_tls *tls_config);
 void find_channel_by_ip_address(struct rasta_handle *h, struct sockaddr_in sender, int *red_channel_idx, int *transport_channel_idx);
