+++ src/c/transport/udp.c
@@ -7,9 +7,9 @@
 #include <string.h> //memset
 #include <unistd.h>
 
+#include "udp.h"
 #include <rasta/bsd_utils.h>
 #include <rasta/rmemory.h>
-#include "udp.h"
 
 #include "transport.h"
 
@@ -18,14 +18,14 @@
 static void handle_tls_mode(rasta_transport_socket *transport_socket) {
     const rasta_config_tls *tls_config = transport_socket->tls_config;
     switch (tls_config->mode) {
-        case TLS_MODE_DISABLED: {
-            transport_socket->tls_mode = TLS_MODE_DISABLED;
-            break;
-        }
-        default: {
-            fprintf(stderr, "Unknown or unsupported TLS mode: %u", tls_config->mode);
-            abort();
-        }
+    case TLS_MODE_DISABLED: {
+        transport_socket->tls_mode = TLS_MODE_DISABLED;
+        break;
+    }
+    default: {
+        fprintf(stderr, "Unknown or unsupported TLS mode: %u", tls_config->mode);
+        abort();
+    }
     }
 }
 
@@ -112,77 +112,4 @@ void udp_init(rasta_transport_socket *transport_socket, const rasta_config_tls *
         abort();
     }
     transport_socket->file_descriptor = file_desc;
-}
-
-void transport_create_socket(struct rasta_handle *h, rasta_transport_socket *socket, int id, const rasta_config_tls *tls_config) {
-    // init and bind sockets
-    socket->id = id;
-    udp_init(socket, tls_config);
-
-    memset(&socket->receive_event, 0, sizeof(fd_event));
-    socket->receive_event.callback = channel_receive_event;
-    socket->receive_event.carry_data = &socket->receive_event_data;
-    socket->receive_event.fd = socket->file_descriptor;
-    
-    memset(&socket->receive_event_data, 0, sizeof(struct receive_event_data));
-    socket->receive_event_data.h = h;
-    socket->receive_event_data.socket = socket;
-    // Iff channel == NULL the receive event operates in 'UDP/DTLS mode'
-    socket->receive_event_data.channel = NULL;
-    socket->receive_event_data.connection = NULL;
-
-    add_fd_event(h->ev_sys, &socket->receive_event, EV_READABLE);
-}
-
-int transport_connect(rasta_transport_socket *socket, rasta_transport_channel *channel, rasta_config_tls tls_config) {
-    UNUSED(tls_config);
-
-    enable_fd_event(&socket->receive_event);
-
-    channel->id = socket->id;
-    channel->tls_mode = socket->tls_mode;
-    channel->file_descriptor = socket->file_descriptor;
-
-    // We can regard UDP channels as 'always connected' (no re-dial possible)
-    channel->connected = true;
-
-    return 0;
-}
-
-int transport_redial(rasta_transport_channel *channel, rasta_transport_socket *socket) {
-    // We can't reconnect when using UDP
-    UNUSED(channel); UNUSED(socket);
-    return -1;
-}
-
-void transport_close(rasta_transport_channel *channel) {
-    UNUSED(channel);
-}
-
-void send_callback(redundancy_mux *mux, struct RastaByteArray data_to_send, rasta_transport_channel *channel, unsigned int channel_index) {
-    UNUSED(mux); UNUSED(channel_index);
-    udp_send(channel, data_to_send.bytes, data_to_send.length, channel->remote_ip_address, channel->remote_port);
-}
-
-ssize_t receive_callback(struct receive_event_data *data, unsigned char *buffer, struct sockaddr_in *sender) {
-    return udp_receive(data->socket, buffer, MAX_DEFER_QUEUE_MSG_SIZE, sender);
-}
-
-void transport_listen(struct rasta_handle *h, rasta_transport_socket *socket) {
-    UNUSED(h);
-    enable_fd_event(&socket->receive_event);
-}
-
-bool transport_bind(struct rasta_handle *h, rasta_transport_socket *socket, const char *ip, uint16_t port) {
-    UNUSED(h);
-    return udp_bind_device(socket, ip, port);
-}
-
-int transport_accept(rasta_transport_socket *socket, struct sockaddr_in *addr) {
-    UNUSED(socket); UNUSED(addr);
-    return 0;
-}
-
-void transport_init(struct rasta_handle *h, rasta_transport_channel* channel, unsigned id, const char *host, uint16_t port, const rasta_config_tls *tls_config) {
-    transport_init_base(h, channel, id, host, port, tls_config);
 }
